requirements.txt
torch
json
Pillow
fastapi
pandas